requirements.txt
# Core Dependencies
streamlit
langgraph
langchain-core
langchain-openai

# API & Data
requests
google-search-results
python-dotenv

# Utilities
diskcache
